requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.7
asyncio==3.4.3
